+++ app/Providers/Filament/PanelUserPanelProvider.php
@@ -4,6 +4,7 @@ namespace App\Providers\Filament;
 
 use App\Enum\Guard\GuardEnum;
 use App\Enum\PanelEnum;
+use App\Filament\Auth\PasskeyProvider;
 use App\Filament\Admin\Pages\Auth\EditProfile;
 use App\Filament\Standard\Pages\Auth\EditTenantProfile;
 use App\Filament\Standard\Pages\Auth\Register;
@@ -152,6 +153,7 @@ class PanelUserPanelProvider extends PanelProvider
             AppAuthentication::make()
                 ->recoverable(),
             EmailAuthentication::make(),
+            PasskeyProvider::make(),
         ]);
     }